[tech] - refactor
